PointCloudRasterizers

PointCloudRasterizers

Example Julia package repo.

Base.reduce — Method.

Reduce multiple points to single value in given indexed pointcloud.

source

PointCloudRasterizers.PointCloudIndex — Type.

GeoArray with the number of points in each cell and an index for each point pointing to the cell its in.

source

Base.filter! — Function.

Filter out index on given condition.

source